1450 Part II - update point properties.
