+++ core/specfem/source/dim3/vector_source/force_source.cpp
@@ -2,6 +2,7 @@
 #include "globals.h"
 #include "source_time_function/interface.hpp"
 #include "specfem/source.hpp"
+#include "specfem/macros.hpp"
 #include "specfem_setup.hpp"
 #include "utilities/interface.hpp"
 #include "yaml-cpp/yaml.h"
