--- docs/sections/api/specfem/point/kernels/dim2/acoustic_isotropic.rst
@@ -1,9 +1,0 @@
-.. _specfem_point_kernels_dim2_acoustic_isotropic:
-
-2D Acoustic Isotropic Point Kernels
-===================================
-
-.. doxygengroup:: specfem_point_kernels_dim2_acoustic_isotropic
-    :members:
-    :private-members:
-    :content-only:
