--- docs/sections/api/specfem/point/kernels/dim2/elastic_isotropic.rst
@@ -1,9 +1,0 @@
-.. _specfem_point_kernels_elastic_isotropic:
-
-Elastic Isotropic Point Kernels
-===============================
-
-.. doxygengroup:: specfem_point_kernels_elastic_isotropic
-    :members:
-    :private-members:
-    :content-only:
