--- docs/sections/api/specfem/point/kernels/dim2/poroelastic_isotropic.rst
@@ -1,9 +1,0 @@
-.. _specfem_point_kernels_dim2_poroelastic_isotropic:
-
-2D Poroelastic Isotropic Point Kernels
-======================================
-
-.. doxygengroup:: specfem_point_kernels_dim2_poroelastic_isotropic
-    :members:
-    :private-members:
-    :content-only:
