--- docs/sections/api/specfem/point/properties/dim2/elastic_anisotropic.rst
@@ -1,9 +1,0 @@
-.. _specfem_point_properties_dim2_elastic_anisotropic:
-
-2D Elastic Anisotropic
-======================
-
-.. doxygengroup:: specfem_point_properties_dim2_elastic_anisotropic
-    :members:
-    :private-members:
-    :content-only:
