--- docs/sections/api/specfem/point/properties/elastic_isotropic.rst
@@ -1,9 +1,0 @@
-.. _specfem_point_properties_elastic_isotropic:
-
-Elastic Isotropic
-=================
-
-.. doxygengroup:: specfem_point_properties_elastic_isotropic
-    :members:
-    :private-members:
-    :content-only:
