+++ include/medium/dim2/acoustic/isotropic/properties.hpp
@@ -1,0 +1,74 @@
+#pragma once
+
+#include "medium/impl/point_container.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem::point::impl::properties {
+
+
+/**
+ * @defgroup specfem_point_properties_dim2_acoustic_isotropic 2D Acoustic
+ * Properties
+ * @{
+ */
+
+/**
+ * @ingroup specfem_point_properties_dim2_acoustic_isotropic
+ * @brief Data container to hold properties of 2D acoustic media at a quadrature
+ * point
+ *
+ * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
+ *
+ * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
+ *
+ * - Get the inverse of density @f$ \frac{1}{\rho} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho_inverse() const @endcode
+ *
+ * - Get the bulk modulus @f$ \kappa @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type kappa() const @endcode
+ *
+ * <b> Additional accessor functions </b>
+ *
+ * - Get the inverse of bulk modulus @f$ \frac{1}{\kappa} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type kappa_inverse() const
+ *   @endcode
+ *
+ * - Get the inverse of the product of density and P-wave velocity @f$
+ * \frac{1}{\rho v_p} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho_vpinverse() const
+ *   @endcode
+ *
+ */
+template <bool UseSIMD>
+struct data_container<specfem::dimension::type::dim2,
+                      specfem::element::medium_tag::acoustic,
+                      specfem::element::property_tag::isotropic, UseSIMD>
+    : PropertyAccessor<specfem::dimension::type::dim2,
+                       specfem::element::medium_tag::acoustic,
+                       specfem::element::property_tag::isotropic, UseSIMD> {
+
+private:
+  using base_type = PropertyAccessor<
+      specfem::dimension::type::dim2, specfem::element::medium_tag::acoustic,
+      specfem::element::property_tag::isotropic, UseSIMD>; ///< Base type of the
+                                                           ///< point properties
+public:
+  using value_type = typename base_type::value_type; ///< Type of the properties
+  using simd = typename base_type::simd;
+
+  POINT_CONTAINER(rho_inverse, kappa)
+
+  KOKKOS_INLINE_FUNCTION const value_type kappa_inverse() const {
+    return (static_cast<type_real>(1.0)) /
+           kappa(); ///< @f$ \frac{1}{\lambda + 2\mu} @f$
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type rho_vpinverse() const {
+    return Kokkos::sqrt(rho_inverse() * kappa_inverse()); ///< @f$ \frac{1}{\rho
+                                                          ///< v_p} @f$
+  }
+};
+
+///@} end of group specfem_point_properties_dim2_acoustic_isotropic
+
+} // namespace specfem::point::impl::properties
