+++ include/medium/dim2/elastic/anisotropic/properties.hpp
@@ -1,0 +1,95 @@
+#pragma once
+
+#include "medium/impl/point_container.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem::point::impl::properties {
+
+/**
+ * @defgroup specfem_point_properties_dim2_elastic_anisotropic 2D Elastic
+ * Anisotropic Properties
+ * @{
+ */
+
+/**
+ * @ingroup specfem_point_properties_dim2_elastic_anisotropic
+ * @brief Data container to hold properties of 2D anisotropic elastic media at a
+ * quadrature point
+ *
+ * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
+ *
+ * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
+ *
+ * - Get stiffness tensor component @f$ c_{11} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c11() const @endcode
+ *
+ * - Get stiffness tensor component @f$ c_{13} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c13() const @endcode
+ *
+ * - Get stiffness tensor component @f$ c_{15} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c15() const @endcode
+ *
+ * - Get stiffness tensor component @f$ c_{33} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c33() const @endcode
+ *
+ * - Get stiffness tensor component @f$ c_{35} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c35() const @endcode
+ *
+ * - Get stiffness tensor component @f$ c_{55} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c55() const @endcode
+ *
+ * - Get stiffness tensor component @f$ c_{12} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c12() const @endcode
+ *
+ * - Get stiffness tensor component @f$ c_{23} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c23() const @endcode
+ *
+ * - Get stiffness tensor component @f$ c_{25} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type c25() const @endcode
+ *
+ * - Get the density @f$ \rho @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho() const @endcode
+ *
+ * <b> Additional accessor functions </b>
+ *
+ * - Compute the product of density and P-wave velocity squared, i.e., @f$ \rho
+ *   v_p^2 = \rho c_{33} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho_vp() const @endcode
+ * - Compute the product of density and S-wave velocity squared, i.e., @f$ \rho
+ *   v_s^2 = \rho c_{55} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho_vs() const @endcode
+ *
+ */
+template <specfem::element::medium_tag MediumTag, bool UseSIMD>
+struct data_container<
+    specfem::dimension::type::dim2, MediumTag,
+    specfem::element::property_tag::anisotropic, UseSIMD,
+    std::enable_if_t<specfem::element::is_elastic<MediumTag>::value> >
+    : public PropertyAccessor<specfem::dimension::type::dim2, MediumTag,
+                              specfem::element::property_tag::anisotropic,
+                              UseSIMD> {
+
+private:
+  using base_type =
+      PropertyAccessor<specfem::dimension::type::dim2, MediumTag,
+                       specfem::element::property_tag::anisotropic,
+                       UseSIMD>; ///< Base type of the
+                                 ///< point properties
+
+public:
+  using value_type = typename base_type::value_type; ///< Type of the properties
+  using simd = typename base_type::simd;
+
+  POINT_CONTAINER(c11, c13, c15, c33, c35, c55, c12, c23, c25, rho)
+
+  KOKKOS_INLINE_FUNCTION const value_type rho_vp() const {
+    return Kokkos::sqrt(rho() * c33()); ///< P-wave velocity @f$ \rho v_p @f$
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type rho_vs() const {
+    return Kokkos::sqrt(rho() * c55()); ///< S-wave velocity @f$ \rho v_s @f$
+  }
+};
+///@} end of group specfem_point_properties_dim2_elastic_anisotropic
+
+} // namespace specfem::point::impl::properties
