+++ include/medium/dim2/electromagnetic/isotropic/properties.hpp
@@ -1,0 +1,62 @@
+#pragma once
+
+#include "medium/impl/point_container.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem::point::impl::properties {
+
+/**
+ * @defgroup specfem_point_properties_dim2_electromagnetic_isotropic 2D
+ * Electromagnetic Isotropic Properties
+ * @{
+ */
+
+/**
+ * @brief Data container to hold properties of 2D electromagnetic media at a
+ * quadrature point
+ *
+ * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
+ *
+ * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
+ *
+ * - Get the inverse of permeability @f$ \frac{1}{\mu_0} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type mu0_inv() const @endcode
+ *
+ * - Get permittivity tensor component @f$ \epsilon_{11} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type eps11() const @endcode
+ *
+ * - Get permittivity tensor component @f$ \epsilon_{33} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type eps33() const @endcode
+ *
+ * - Get conductivity tensor component @f$ \sigma_{11} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type sig11() const @endcode
+ *
+ * - Get conductivity tensor component @f$ \sigma_{33} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type sig33() const @endcode
+ */
+template <specfem::element::medium_tag MediumTag, bool UseSIMD>
+struct data_container<
+    specfem::dimension::type::dim2, MediumTag,
+    specfem::element::property_tag::isotropic, UseSIMD,
+    std::enable_if_t<specfem::element::is_electromagnetic<MediumTag>::value> >
+    : public PropertyAccessor<specfem::dimension::type::dim2, MediumTag,
+                              specfem::element::property_tag::isotropic,
+                              UseSIMD> {
+
+private:
+  using base_type = PropertyAccessor<specfem::dimension::type::dim2, MediumTag,
+                                     specfem::element::property_tag::isotropic,
+                                     UseSIMD>; ///< Base type of the
+                                               ///< point properties
+
+public:
+  using value_type = typename base_type::value_type; ///< Type of the properties
+
+  using simd = typename base_type::simd;
+
+  POINT_CONTAINER(mu0_inv, eps11, eps33, sig11, sig33)
+};
+
+/** @} end of group specfem_point_properties_dim2_electromagnetic_isotropic */
+
+} // namespace specfem::point::impl::properties
