+++ include/medium/dim2/poroelastic/isotropic/properties.hpp
@@ -1,0 +1,193 @@
+#pragma once
+
+#include "medium/impl/point_container.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem::point::impl::properties {
+
+/**
+ * @defgroup specfem_point_properties_dim2_poroelastic_isotropic 2D Poroelastic
+ * Isotropic Properties
+ * @{
+ */
+
+/**
+ * @ingroup specfem_point_properties_dim2_poroelastic_isotropic
+ * @brief Data container to hold properties of 2D poroelastic media at a
+ * quadrature point
+ *
+ * @tparam UseSIMD Boolean indicating whether to use SIMD intrinsics
+ *
+ * <b> Accessor functions generated by the </b> \c POINT_CONTAINER \c MACRO
+ *
+ * - Get porosity @f$ \phi @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type phi() const @endcode
+ *
+ * - Get solid density @f$ \rho_s @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho_s() const @endcode
+ *
+ * - Get fluid density @f$ \rho_f @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho_f() const @endcode
+ *
+ * - Get tortuosity @f$ \tau @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type tortuosity() const @endcode
+ *
+ * - Get shear modulus @f$ \mu_G @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type mu_G() const @endcode
+ *
+ * - Get Biot's modulus @f$ H_Biot @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type H_Biot() const @endcode
+ *
+ * - Get Biot's modulus @f$ C_Biot @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type C_Biot() const @endcode
+ *
+ * - Get Biot's modulus @f$ M_Biot @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type M_Biot() const @endcode
+ *
+ * - Get permeability tensor component @f$ k_{xx} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type permxx() const @endcode
+ *
+ * - Get permeability tensor component @f$ k_{xz} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type permxz() const @endcode
+ *
+ * - Get permeability tensor component @f$ k_{zz} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type permzz() const @endcode
+ *
+ * - Get fluid viscosity @f$ \eta_f @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type eta_f() const @endcode
+ *
+ * <b> Additional accessor functions </b>
+ *
+ * - Get Lame's parameter @f$ \lambda_G @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type lambda_G() const @endcode
+ *
+ * - Get Lame's parameter @f$ \lambda + 2\mu_G @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type lambdaplus2mu_G() const
+ *   @endcode
+ *
+ * - Get the inverse of permeability tensor component @f$ k_{xx} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type inverse_permxx() const
+ *   @endcode
+ *
+ * - Get the inverse of permeability tensor component @f$ k_{xz} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type inverse_permxz() const
+ *   @endcode
+ *
+ * - Get the inverse of permeability tensor component @f$ k_{zz} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type inverse_permzz() const
+ *   @endcode
+ *
+ * - Get the average density @f$ \rho_{bar} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type rho_bar() const @endcode
+ *
+ * - Get the P-wave velocity @f$ v_{pI} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type vpI() const @endcode
+ *
+ * - Get the P-wave velocity @f$ v_{pII} @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type vpII() const @endcode
+ *
+ * - Get the S-wave velocity @f$ v_s @f$
+ *   @code KOKKOS_INLINE_FUNCTION const value_type vs() const @endcode
+ */
+template <bool UseSIMD>
+struct data_container<specfem::dimension::type::dim2,
+                      specfem::element::medium_tag::poroelastic,
+                      specfem::element::property_tag::isotropic, UseSIMD>
+    : public PropertyAccessor<specfem::dimension::type::dim2,
+                              specfem::element::medium_tag::poroelastic,
+                              specfem::element::property_tag::isotropic,
+                              UseSIMD> {
+
+private:
+  using base_type = PropertyAccessor<
+      specfem::dimension::type::dim2, specfem::element::medium_tag::poroelastic,
+      specfem::element::property_tag::isotropic, UseSIMD>; ///< Base type of the
+                                                           ///< point properties
+
+public:
+  using value_type = typename base_type::value_type; ///< Type of the properties
+
+  using simd = typename base_type::simd;
+
+  POINT_CONTAINER(phi, rho_s, rho_f, tortuosity, mu_G, H_Biot, C_Biot, M_Biot,
+                  permxx, permxz, permzz, eta_f)
+
+  KOKKOS_INLINE_FUNCTION const value_type lambda_G() const {
+    return H_Biot() - (static_cast<type_real>(2.0)) * mu_G();
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type lambdaplus2mu_G() const {
+    return lambda_G() + (static_cast<type_real>(2.0)) * mu_G();
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type inverse_permxx() const {
+    const value_type determinant =
+        permxx() * permzz() - permxz() * permxz(); ///< determinant of the
+                                                   ///< permeability tensor
+    return permzz() / determinant; ///< inverse of the permeability tensor
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type inverse_permxz() const {
+    const value_type determinant =
+        permxx() * permzz() - permxz() * permxz(); ///< determinant of the
+                                                   ///< permeability tensor
+    return static_cast<type_real>(-1.0) * permxz() /
+           determinant; ///< inverse of the permeability tensor
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type inverse_permzz() const {
+    const value_type determinant =
+        permxx() * permzz() - permxz() * permxz(); ///< determinant of the
+                                                   ///< permeability tensor
+    return permxx() / determinant; ///< inverse of the permeability tensor
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type rho_bar() const {
+    return (((static_cast<type_real>(1.0) - this->phi()) * this->rho_s()) +
+            (this->phi() * this->rho_f()));
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type vpI() const {
+    ///< Helper variable for readability
+    const auto phi_over_tort = this->phi() / this->tortuosity();
+    const auto afactor = rho_bar() - phi_over_tort * rho_f();
+    const auto bfactor =
+        this->H_Biot() + phi_over_tort * rho_bar() / rho_f() * this->M_Biot() -
+        static_cast<type_real>(2.0) * phi_over_tort * this->C_Biot();
+    const auto cfactor =
+        phi_over_tort / rho_f() *
+        (this->H_Biot() * this->M_Biot() - this->C_Biot() * this->C_Biot());
+
+    return Kokkos::sqrt(bfactor + Kokkos::sqrt(bfactor * bfactor -
+                                               static_cast<type_real>(4.0) *
+                                                   afactor * cfactor)) /
+           (static_cast<type_real>(2.0) * afactor);
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type vpII() const {
+    ///< Helper variable for readability
+    const auto phi_over_tort = this->phi() / this->tortuosity();
+    const auto afactor = rho_bar() - phi_over_tort * rho_f();
+    const auto bfactor =
+        this->H_Biot() + phi_over_tort * rho_bar() / rho_f() * this->M_Biot() -
+        static_cast<type_real>(2.0) * phi_over_tort * this->C_Biot();
+    const auto cfactor =
+        phi_over_tort / rho_f() *
+        (this->H_Biot() * this->M_Biot() - this->C_Biot() * this->C_Biot());
+
+    return Kokkos::sqrt(bfactor - Kokkos::sqrt(bfactor * bfactor -
+                                               static_cast<type_real>(4.0) *
+                                                   afactor * cfactor)) /
+           (static_cast<type_real>(2.0) * afactor);
+  }
+
+  KOKKOS_INLINE_FUNCTION const value_type vs() const {
+    ///< Helper variable for readability
+    const auto phi_over_tort = this->phi() / this->tortuosity();
+    const auto afactor = rho_bar() - phi_over_tort * rho_f();
+    return Kokkos::sqrt(mu_G() / afactor);
+  }
+};
+/**@} end of group specfem_point_properties_dim2_poroelastic_isotropic */
+
+} // namespace specfem::point::impl::properties
